refactor: let table filters inside alerts table
